Moving code from program.cs into nimrod.csproj and introducing abstraction for the file system.
