+++ Nimrod.Console/packages.config
@@ -1,4 +1,5 @@
 ﻿<?xml version="1.0" encoding="utf-8"?>
 <packages>
   <package id="CommandLineParser" version="1.9.71" targetFramework="net4" />
+  <package id="System.IO.Abstractions" version="2.0.0.124" targetFramework="net4" />
 </packages>
