+++ Nimrod.Test/Nimrod.Test.csproj
@@ -107,7 +107,6 @@
     <Compile Include="TypeDiscoveryTests.cs" />
     <Compile Include="TypeExtensionsTest.cs" />
     <Compile Include="Writers\BaseWriterTests.cs" />
-    <Compile Include="Writers\ControllerWriterTests.cs" />
     <Compile Include="Writers\EnumWriterTest.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="ToTypescriptTests.cs" />
